added new prefabs and modified scripts i pushed last

+++ Assets/Scripts/simpleObjectSpawner.cs
@@ -7,14 +7,14 @@ public class simpleObjectSpawner : MonoBehaviour
     [SerializeField] Transform[] spawnPos;
     [SerializeField] GameObject objectToSpawn;
 
-    [SerializeField] int startSpawnTime;
+    [SerializeField] float startSpawnTime;
     [SerializeField] float startNewSpawnTime;
     
 
     void Start()
     {
         // Call the Spawn function after a delay of the spawnTime and then continue to call after the same amount of time.
-        InvokeRepeating("Spawn", startSpawnTime, startNewSpawnTime);
+            InvokeRepeating("Spawn", startSpawnTime, startNewSpawnTime);
     }
 
     void Update()
